Add files via upload
This program makes use of PyPDF2 library to parse through the student handbooks of MISM and MSPPM programs, to extract the list of electives for each. It then uses RegEx to pattern-match and populates the course code, description and number of units of each elective into a dataframe. Each program uses a separate function, owing the difference in patterns of the handbooks.